+++ src/js/pages/app.page.js
@@ -1,6 +1,4 @@
 import React from 'react'
-import LoadingComponent from '../components/loading.component'
-import ErrorComponent from '../components/error.component'
 import { connect } from 'react-redux'
 import { Router, Route, Link, withRouter, Switch } from 'react-router-dom'
 import { browserHistory } from '../services/browser-history.service'
@@ -8,14 +6,12 @@ import AuthService from '../services/auth.service'
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { initialize, fetchUser } from '../actions'
-import RoomsPartial from '../partials/rooms.partial'
-import RoomPartial from '../partials/room.partial'
-import DockPartial from '../partials/dock.partial'
-import ToolbarPartial from '../partials/toolbar.partial'
-import MembersPartial from '../partials/members.partial'
 import GraphqlService from '../services/graphql.service'
-import NotificationComponent from '../components/notification.component'
 import CookieService from '../services/cookies.service'
+import { Avatar, Loading, Error, Notification } from '@weekday/elements'
+import RoomsComponent from '../components/rooms.component'
+import RoomComponent from '../components/room.component'
+import DockComponent from '../components/dock.component'
 
 const App = styled.div`
   background-color: white;
@@ -105,15 +101,15 @@ class AppPage extends React.Component {
 
   // prettier-ignore
   render() {
-    if (!this.props.common.user.id) return <LoadingComponent show={true} />
+    if (!this.props.common.user.id) return <Loading show={true} />
 
     return (
       <App className="column align-items-start app-page">
-        <LoadingComponent show={this.props.common.loading} />
-        <ErrorComponent message={this.props.common.error} />
+        <Loading show={this.props.common.loading} />
+        <Error message={this.props.common.error} />
 
         {this.state.pushNotifications &&
-          <NotificationComponent
+          <Notification
             text="Push notifications are disabled. You will not receive any notifications in the background."
             actionText={null}
             onDismissClick={this.dismissPushNotifications}
@@ -123,11 +119,9 @@ class AppPage extends React.Component {
 
         <Router history={browserHistory}>
           <div className="row w-100 h-100 align-items-start align-content-start justify-content-start flex-1">
-            <Route path="/app" component={DockPartial} />
-            <Route path="/app/team/:teamId" component={RoomsPartial} />
-            <Route path="/app/team/:teamId/room/:roomId" component={RoomPartial} />
-            <Route path="/app/team/:teamId/room/:roomId/members" component={MembersPartial} />
-            <Route path="/app/team/:teamId/room/:roomId" component={ToolbarPartial} />
+            <Route path="/app" component={DockComponent} />
+            <Route path="/app/team/:teamId" component={RoomsComponent} />
+            <Route path="/app/team/:teamId/room/:roomId" component={RoomComponent} />
           </div>
         </Router>
       </App>
