+++ src/js/pages/auth.page.js
@@ -1,7 +1,5 @@
 import React from 'react'
 import AuthService from '../services/auth.service'
-import LoadingComponent from '../components/loading.component'
-import ErrorComponent from '../components/error.component'
 import { connect } from 'react-redux'
 import GraphqlService from '../services/graphql.service'
 import styled from 'styled-components'
@@ -9,7 +7,7 @@ import { Formik } from 'formik'
 import * as Yup from 'yup'
 import PropTypes from 'prop-types'
 import { fetchUser } from '../actions'
-import { Button } from '@weekday/elements'
+import { Loading, Error, Button, Input, Textarea } from '@weekday/elements'
 
 const Auth = styled.div`
   height: 100%;
@@ -86,42 +84,6 @@ const Footer = styled.div`
 const Avatar = styled.div`
   padding: 50px;
   width: 100%;
-`
-
-const Textarea = styled.textarea`
-  color: #202529;
-  font-size: 16px;
-  font-weight: 400;
-  border: none;
-  border-top: 1px solid #f1f3f5;
-  padding: 20px;
-  width: 100%;
-  text-align: center;
-  resize: none;
-
-  &::placeholder {
-    color: #ebedef;
-  }
-`
-
-const Input = styled.input`
-  color: #202529;
-  font-size: 30px;
-  font-weight: 400;
-  padding: 20px;
-  width: 100%;
-  text-align: center;
-  border: none;
-  border-top: 1px solid #f1f3f5;
-
-  &.error {
-    border-top: 1px solid red;
-    border-bottom: 1px solid red;
-  }
-
-  &::placeholder {
-    color: #ebedef;
-  }
 `
 
 const Usernames = styled.div`
@@ -346,8 +308,8 @@ class AuthPage extends React.Component {
             <LogoText>weekday</LogoText>
           </Logo>
 
-          <LoadingComponent show={this.state.loading} />
-          <ErrorComponent message={this.state.error} />
+          <Loading show={this.state.loading} />
+          <Error message={this.state.error} />
 
           {this.state.view == "password" &&
             <React.Fragment>
